+++ src/main/java/brot/lwjgl/game/scene/collision/AabbObject.java
@@ -2,21 +2,22 @@ package brot.lwjgl.game.scene.collision;
 
 import brot.lwjgl.engine.scene.entity.AABB;
 import com.google.common.base.Preconditions;
+import org.joml.Vector2d;
 import org.joml.Vector2f;
 
 public class AabbObject implements AABB {
-    public Vector2f centerPoint;
-    public Vector2f halfExtents;
+    public Vector2d centerPoint;
+    public Vector2d halfExtents;
 
     public AabbObject() {
         this(null, null);
     }
 
-    public AabbObject(Vector2f centre, Vector2f halfExtents) {
+    public AabbObject(Vector2d centre, Vector2d halfExtents) {
         initialise(centre, halfExtents);
     }
 
-    public void initialise(Vector2f center, Vector2f halfExtents) {
+    public void initialise(Vector2d center, Vector2d halfExtents) {
         this.centerPoint = center;
         this.halfExtents = halfExtents;
         if (halfExtents != null) {
@@ -25,12 +26,12 @@ public class AabbObject implements AABB {
     }
 
     @Override
-    public Vector2f getCenter() {
+    public Vector2d getCenter() {
         return centerPoint;
     }
 
     @Override
-    public Vector2f getHalfExtents() {
+    public Vector2d getHalfExtents() {
         return halfExtents;
     }
